Adds fix for issue #1: terminal size error
This imports a pythno standard package `os`, which is used to check the
current terminal size. This is then compared to the minimum required
size for multicell to run. If the terminal is too small a MulticellError
is raised.

Note: Currently the minimum required size are given by constants at the
begining of the program. It would be better if they were calculated
explitly from the code.